Didn't add any of the offset maths to the final tile-to-check-builder, so all of my fixes had no a/e?ffect.
However, I'm now getting a smooth 60fps after a couple of days of 20.

Only neighbouring tiles are considered for collision checks, using the tile class for some helpful functions. This would have needed to be done at some point anyway, so this wasn't all a waste of time.

It also led to me adding some helpful debug features, such as a grid showing tiles and the FPS counter.